Convolution now deals with doubles and returns a uint32 matrix.

+++ convolution.m
@@ -1,8 +1,8 @@
 function x = convolution(f,g)
-  x = f;
+  x = uint32(f);
   for i = 2:size(f)(1)-1
     for j = 2:size(f)(2)-1
-    x(i, j) = sum(sum( f(i-1:i+1, j-1:j+1) .* g));
+    x(i, j) = uint32(sum(sum( double(f(i-1:i+1, j-1:j+1)) .* double(g))));
     endfor
   endfor
 endfunction
